+++ src/features/filters/queries/segments/limitByPatchAffected.ts
@@ -1,0 +1,7 @@
+import { Prisma } from '@prisma/client'
+
+export function limitByPatchAffected(includePatchAffectedBuilds: boolean) {
+  return includePatchAffectedBuilds
+    ? Prisma.empty
+    : Prisma.sql`AND Build.isPatchAffected=false`
+}