+++ src/helpful-queries/patch_2024-02-29.sql
@@ -1,0 +1,77 @@
+# Flag builds that contain Spectral Blade + Resonance
+UPDATE
+  Build
+SET
+  isPatchAffected = true
+WHERE
+  id IN (
+    SELECT
+      id
+    FROM
+      (
+        SELECT
+          Build.id
+        FROM
+          Build
+          JOIN BuildItems ON Build.id = BuildItems.buildId
+        WHERE
+          BuildItems.itemId IN ('l3zte1', 'dmizlm')
+          AND Build.isPublic = true
+        GROUP BY
+          Build.id
+        HAVING
+          COUNT(DISTINCT BuildItems.itemId) = 2
+      ) AS tmp
+  );
+
+# Flag builds that contain Stone of Malevolence
+UPDATE
+  Build
+SET
+  isPatchAffected = true
+WHERE
+  id IN (
+    SELECT
+      id
+    FROM
+      (
+        SELECT
+          Build.id
+        FROM
+          Build
+          JOIN BuildItems ON Build.id = BuildItems.buildId
+        WHERE
+          BuildItems.itemId IN ('kx65km')
+          AND Build.isPublic = true
+        GROUP BY
+          Build.id
+        HAVING
+          COUNT(DISTINCT BuildItems.itemId) = 1
+      ) AS tmp
+  );
+
+# Flag builds that contain Sequenced Shot
+UPDATE
+  Build
+SET
+  isPatchAffected = true
+WHERE
+  id IN (
+    SELECT
+      id
+    FROM
+      (
+        SELECT
+          Build.id
+        FROM
+          Build
+          JOIN BuildItems ON Build.id = BuildItems.buildId
+        WHERE
+          BuildItems.itemId IN ('23ztdj')
+          AND Build.isPublic = true
+        GROUP BY
+          Build.id
+        HAVING
+          COUNT(DISTINCT BuildItems.itemId) = 1
+      ) AS tmp
+  );